requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
exceptiongroup==1.2.0
flatbuffers==23.5.26
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
MLB-StatsAPI==1.7.1
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
pandas==2.2.1
platformdirs==4.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
zope.interface==6.2
